xpander-comparison/requirement.txt
xpander-sdk
openai
python-dotenv
pyyaml
colorama
requests
colorlog
pandas
jupyter
seaborn
matplotlib
